Renamed writer to formatter
